Fixed potential scope issues for $php_errormsg
Simply replacing any usage of $php_errormsg (as this has been removed in
PHP8.0) with error_get_last() could potentially lead to fetching an
error that has nothing to do with the code we actually want to inspect.
Thus error_clear_last() is called before executing the lines prone to
cause an error. As error_clear_last() is PHP7+ Symfony's polyfill was
added as a dependency for backward compatbility.
The return value of error_get_last() is checked in accordance with the
recommendation of the polyfill.